refactor : 식당 상세정보 즐겨찾기 이벤트 핸들러 storeDetail.js로 이동
