+++ src/render/storeRenderer.js
@@ -1,8 +1,9 @@
-import Store from "../components/Store.js";
+import Store from "../components/store.js";
 import IMG_SRC from "../constants/imgSrc.js";
 import storage from "../utils/storage.ts";
 import validate from "../utils/validate.ts";
 import modalRenderer from "./modalRenderer.js";
+import { v4 as uuidv4 } from "uuid";
 
 const storeRenderer = {
   // 새로운 식당 추가
